Vector: refactor of stack header
Add Stack on vector implementation. The idea is simple: Define Vector to
be Stack of elements of same type, Define VECTOR_PUSH_BACK to be
STACK_PUSH and VECTOR_REMOVE(VECTOR_SIZE-1) to be SATCK_POP.

So that, we have a stack wich could be nativelly converted to vector in
purpose to access its bottom.